Initial working SQLite3 db
Schema file and sample database with that schema loaded. The collatz.sh
script using a different db filename that is not being added to the
repo, and in fact is listed in the (newly added) .gitignore file, b/c
you shouldn't version control output data with your git repo. silly.