helper script to compare databases
Use SQL queries to extract batches of rows and compare the batches by
piping through md5sum hash, just as a quick check for any differences.
Ideally if I'm trying to do the same thing a new iteration of the
database/code, then the hashes will be the same for all batches
analyzed.

Sample output:
  $ ./compare-collatz-tables.bash
  Batch 0: range 0 to 100
  31a0f8a16b04e96c59aa219480c19234  - data/collatz.01.db
  31a0f8a16b04e96c59aa219480c19234  - data/collatz.02.db
  31a0f8a16b04e96c59aa219480c19234  - data/collatz.03.db

  Batch 1: range 100 to 200
  d41d8cd98f00b204e9800998ecf8427e  - data/collatz.01.db
  d41d8cd98f00b204e9800998ecf8427e  - data/collatz.02.db
  d41d8cd98f00b204e9800998ecf8427e  - data/collatz.03.db

  Batch 2: range 200 to 300
  d41d8cd98f00b204e9800998ecf8427e  - data/collatz.01.db
  d41d8cd98f00b204e9800998ecf8427e  - data/collatz.02.db
  d41d8cd98f00b204e9800998ecf8427e  - data/collatz.03.db