Keep track of calculation time for a Collatz orbit
Shortcuts get calcTime saved and orbits that run into a shortcut get an
estimated calcTime based on the save shortcut calcTime. Also, any new
shortcuts encountered before the end of the orbit they are first noticed
get an estimated calcTime themselves, based on the same shortcut
calcTime encountered/used for the Collatz orbit that was being computed
when the new shortcut(s) was found.