compute distribution of differences in a Collatz orbit
The only thing I "discovered" thus far is that in the distribution for
"consecutive differences" there is always a single instance for each
difference encountered. But this is not very interesting since the
difference between n and collatz(n) is an isomorphism on the natural
numbers (i.e., the diff is -n/2 if n is even and (n+1)/2 if n is odd).

e.g.
Consecutive Diffs Distribution for 5
diff  freq
----  ----
  -4     1 *
  -2     1 *
  -1     1 *
   3     1 *

Consecutive Diffs Distribution for 6
diff  freq
----  ----
  -4     1 *
  -3     1 *
  -2     1 *
  -1     1 *
   2     1 *
   3     1 *

Consecutive Diffs Distribution for 7
diff  freq
----  ----
 -13     1 *
 -10     1 *
  -5     1 *
  -4     1 *
  -2     1 *
  -1     1 *
   3     1 *
   4     1 *
   6     1 *
   7     1 *
   9     1 *