compute distribution of differences in a Collatz orbit
It looks the distribution for "consecutive differences" is always a single
instance for each difference encountered. Interesting. I am not sure if this
is a known or trivial thing to prove.

e.g.
Consecutive Diffs Distribution for 5
diff  freq
----  ----
  -4     1 *
  -2     1 *
  -1     1 *
   3     1 *

Consecutive Diffs Distribution for 6
diff  freq
----  ----
  -4     1 *
  -3     1 *
  -2     1 *
  -1     1 *
   2     1 *
   3     1 *

Consecutive Diffs Distribution for 7
diff  freq
----  ----
 -13     1 *
 -10     1 *
  -5     1 *
  -4     1 *
  -2     1 *
  -1     1 *
   3     1 *
   4     1 *
   6     1 *
   7     1 *
   9     1 *