Build the next start from previous starts and their path lengths
Use sequencial start to build up new starts by adding 2 raised to the
power of
  (bit length of the previous start) + (path length of the previous start) - 1
In the code this is the line
   n = n + (1 << (n.bit_length() + nPathLen - 1))
This is just adding a single '1' to the binary of the previous start,
but placing this bit at a place-value high enough to guarantee that it
will have no effect on the Collatz process until the entire path
traversed by/for the previous start value has been repeated.

For example, with the given initial 'start' value of 3 in this commit,
the first 3 starts are
Decimal:
    3
  131
Binary:
        11
  10000011

The pathLen of 3 is 6, so the padded binary number 0000011 reduces to 1
without encounting a value that uses
  8 = (2 = bit length of '3') + (6 = pathLen of '3')
bits (otherwise the pathLen would need to be at least 7 = (8 - 1). So we
can safely add (2 ** 7) to the start value '3'.

Anyways, this idea needs some tweaks, if it isn't entirely garbage.
Maybe something to ensure that the path of one start value has some
effect on the follow-on start value. The current code is basically
creating a number that is guaranteed to have not collatz orbit affect
for the initial "pathLen of the previous start value" steps, and then
the part of the number this is "new" is just a power of 3, shift 1 or
two bits to the left with "1" or "10" inserted into the first two bits.
It is silly. Damned silly.